fix: validation overwrite bug
